+++ src/render/apply/shell.rs
@@ -15,7 +15,7 @@ impl Sh {
 }
 
 impl Action for Sh {
-    fn execute(&self, mut task: Task, variables: Variables)
+    fn execute(&self, task: &mut Task, variables: &Variables)
         -> Result<(), Error>
     {
         let mut cmd = Command::new("sh");
@@ -27,14 +27,10 @@ impl Action for Sh {
             try!(task.log.redirect_command(&mut cmd));
             cmd.status()
             .map_err(|e| {
-                task.log.log(format_args!(
-                    "Sh {:#?} failed to start: {}\n", cmd, e));
-                Error::CantRun(
-                    task.runner.to_string(), format!("{:#?}", cmd), e)
+                task.log.log_err(format_args!(
+                    "Sh {:#?} failed to start: {}\n", cmd, e))
             }).and_then(|s| if s.success() { Ok(()) } else {
-                task.log.log(format_args!("Sh {:#?}: {}\n", cmd, s));
-                Err(Error::Command(
-                    task.runner.to_string(), format!("{:#?}", cmd), s))
+                Err(task.log.log_err(format_args!("Sh {:#?}: {}\n", cmd, s)))
             })
         } else {
             Ok(())