+++ src/render/renderfile.rs
@@ -60,12 +60,13 @@ quick_error! {
 }
 
 
-fn command_validator<'x>() -> V::Enum<'x> {
+pub fn command_validator<'x>() -> V::Enum<'x> {
     V::Enum::new()
     .option("RootCommand", apply::root_command::RootCommand::config())
     .option("Cmd", apply::cmd::Cmd::config())
     .option("Sh", apply::shell::Sh::config())
     .option("Copy", apply::copy::Copy::config())
+    .option("Condition", apply::condition::Condition::config())
     .option("SplitText", apply::split_text::SplitText::config())
     .option("PeekLog", apply::peek_log::PeekLog::config())
 }